Remove daemonize mode, switch to debug mode instead
In the previous version of this code, the default_value of options[:daemonize] was set to true.  Whenever this flag was set, options[:daemonize] was also evaluated to true.  This meant that you could never start the server without daemonizing.

Renaming the flag to --no-daemonize would have given: -d, --[no-]no-daemonize - Daemonize, so I've renamed this to debug.  The debug flag behaves exactly as you'd expect: -d stops the daemonizing, and prints the output to the console.