Preserve newline when displaying check summaries
Check summaries often have newlines that are significant. For example,
Nagios, or Sensu, plugin output is line-oriented. It would be useful to
obey these newlines when displaying the check output. We do not
want to just use <pre> because that prevents line wrapping. We use
"white-space: pre-line" instead, which preserves all the newlines
in the summary, and adds additional ones by wrapping when necessary to
fit in the enclosing UI elements.

In email templates, it seems that only the <pre> tag works reliably.